refactor: type options directory
