+++ frontend/rust-lib/flowy-grid/src/services/field/mod.rs
@@ -1,5 +1,4 @@
 mod field_builder;
-pub mod select_option;
 pub(crate) mod type_options;
 
 pub use field_builder::*;
