+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/checkbox_type_option/mod.rs
@@ -1,0 +1,6 @@
+mod checkbox_option;
+mod checkbox_option_entities;
+mod tests;
+
+pub use checkbox_option::*;
+pub use checkbox_option_entities::*;
