+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/mod.rs
@@ -1,17 +1,14 @@
-mod checkbox_type_option;
-mod date_type_option;
-mod multi_select_type_option;
-mod number_type_option;
-mod single_select_type_option;
-mod text_type_option;
-mod url_type_option;
+pub mod checkbox_type_option;
+pub mod date_type_option;
+pub mod number_type_option;
+pub mod selection_type_option;
+pub mod text_type_option;
+pub mod url_type_option;
 mod util;
 
 pub use checkbox_type_option::*;
 pub use date_type_option::*;
-pub use multi_select_type_option::*;
-pub use multi_select_type_option::*;
 pub use number_type_option::*;
-pub use single_select_type_option::*;
+pub use selection_type_option::*;
 pub use text_type_option::*;
 pub use url_type_option::*;
