+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/number_type_option/mod.rs
@@ -1,6 +1,9 @@
 #![allow(clippy::module_inception)]
 mod format;
-mod number_type_option;
+mod number_option;
+mod number_option_entities;
+mod tests;
 
 pub use format::*;
-pub use number_type_option::*;
+pub use number_option::*;
+pub use number_option_entities::*;
