+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/selection_type_option/mod.rs
@@ -1,0 +1,7 @@
+mod multi_select_type_option;
+mod select_option;
+mod single_select_type_option;
+
+pub use multi_select_type_option::*;
+pub use select_option::*;
+pub use single_select_type_option::*;
