+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/selection_type_option/single_select_type_option.rs
@@ -1,35 +1,33 @@
 use crate::entities::FieldType;
-
 use crate::impl_type_option;
 use crate::services::cell::{CellBytes, CellData, CellDataChangeset, CellDataOperation, CellDisplayable};
-use crate::services::field::select_option::{
+use crate::services::field::{
     make_selected_select_options, SelectOption, SelectOptionCellChangeset, SelectOptionCellData, SelectOptionIds,
-    SelectOptionOperation, SELECTION_IDS_SEPARATOR,
+    SelectOptionOperation,
 };
-use crate::services::field::type_options::util::get_cell_data;
 use crate::services::field::{BoxTypeOptionBuilder, TypeOptionBuilder};
 use bytes::Bytes;
 use flowy_derive::ProtoBuf;
 use flowy_error::{FlowyError, FlowyResult};
-
 use flowy_grid_data_model::revision::{CellRevision, FieldRevision, TypeOptionDataDeserializer, TypeOptionDataEntry};
-
 use serde::{Deserialize, Serialize};
 
-// Multiple select
+// Single select
 #[derive(Clone, Debug, Default, Serialize, Deserialize, ProtoBuf)]
-pub struct MultiSelectTypeOption {
+pub struct SingleSelectTypeOption {
     #[pb(index = 1)]
     pub options: Vec<SelectOption>,
 
     #[pb(index = 2)]
     pub disable_color: bool,
 }
-impl_type_option!(MultiSelectTypeOption, FieldType::MultiSelect);
+impl_type_option!(SingleSelectTypeOption, FieldType::SingleSelect);
 
-impl SelectOptionOperation for MultiSelectTypeOption {
+impl SelectOptionOperation for SingleSelectTypeOption {
     fn selected_select_option(&self, cell_data: CellData<SelectOptionIds>) -> SelectOptionCellData {
-        let select_options = make_selected_select_options(cell_data, &self.options);
+        let mut select_options = make_selected_select_options(cell_data, &self.options);
+        // only keep option in single select
+        select_options.truncate(1);
         SelectOptionCellData {
             options: self.options.clone(),
             select_options,
@@ -45,7 +43,7 @@ impl SelectOptionOperation for MultiSelectTypeOption {
     }
 }
 
-impl CellDataOperation<SelectOptionIds, SelectOptionCellChangeset> for MultiSelectTypeOption {
+impl CellDataOperation<SelectOptionIds, SelectOptionCellChangeset> for SingleSelectTypeOption {
     fn decode_cell_data(
         &self,
         cell_data: CellData<SelectOptionIds>,
@@ -56,41 +54,22 @@ impl CellDataOperation<SelectOptionIds, SelectOptionCellChangeset> for MultiSele
             return Ok(CellBytes::default());
         }
 
-        let cell_data = self.display_data(cell_data, decoded_field_type, field_rev)?;
-        CellBytes::from(cell_data)
+        self.display_data(cell_data, decoded_field_type, field_rev)
     }
 
     fn apply_changeset(
         &self,
         changeset: CellDataChangeset<SelectOptionCellChangeset>,
-        cell_rev: Option<CellRevision>,
+        _cell_rev: Option<CellRevision>,
     ) -> Result<String, FlowyError> {
-        let content_changeset = changeset.try_into_inner()?;
+        let select_option_changeset = changeset.try_into_inner()?;
         let new_cell_data: String;
-        match cell_rev {
-            None => {
-                new_cell_data = content_changeset.insert_option_id.unwrap_or_else(|| "".to_owned());
-            }
-            Some(cell_rev) => {
-                let cell_data = get_cell_data(&cell_rev);
-                let mut select_ids: SelectOptionIds = cell_data.into();
-                if let Some(insert_option_id) = content_changeset.insert_option_id {
-                    tracing::trace!("Insert multi select option: {}", &insert_option_id);
-                    if select_ids.contains(&insert_option_id) {
-                        select_ids.retain(|id| id != &insert_option_id);
-                    } else {
-                        select_ids.push(insert_option_id);
-                    }
-                }
-
-                if let Some(delete_option_id) = content_changeset.delete_option_id {
-                    tracing::trace!("Delete multi select option: {}", &delete_option_id);
-                    select_ids.retain(|id| id != &delete_option_id);
-                }
-
-                new_cell_data = select_ids.join(SELECTION_IDS_SEPARATOR);
-                tracing::trace!("Multi select cell data: {}", &new_cell_data);
-            }
+        if let Some(insert_option_id) = select_option_changeset.insert_option_id {
+            tracing::trace!("Insert single select option: {}", &insert_option_id);
+            new_cell_data = insert_option_id;
+        } else {
+            tracing::trace!("Delete single select option");
+            new_cell_data = "".to_string()
         }
 
         Ok(new_cell_data)
@@ -98,84 +77,82 @@ impl CellDataOperation<SelectOptionIds, SelectOptionCellChangeset> for MultiSele
 }
 
 #[derive(Default)]
-pub struct MultiSelectTypeOptionBuilder(MultiSelectTypeOption);
-impl_into_box_type_option_builder!(MultiSelectTypeOptionBuilder);
-impl_builder_from_json_str_and_from_bytes!(MultiSelectTypeOptionBuilder, MultiSelectTypeOption);
-impl MultiSelectTypeOptionBuilder {
+pub struct SingleSelectTypeOptionBuilder(SingleSelectTypeOption);
+impl_into_box_type_option_builder!(SingleSelectTypeOptionBuilder);
+impl_builder_from_json_str_and_from_bytes!(SingleSelectTypeOptionBuilder, SingleSelectTypeOption);
+
+impl SingleSelectTypeOptionBuilder {
     pub fn option(mut self, opt: SelectOption) -> Self {
         self.0.options.push(opt);
         self
     }
 }
 
-impl TypeOptionBuilder for MultiSelectTypeOptionBuilder {
+impl TypeOptionBuilder for SingleSelectTypeOptionBuilder {
     fn field_type(&self) -> FieldType {
-        FieldType::MultiSelect
+        FieldType::SingleSelect
     }
 
     fn entry(&self) -> &dyn TypeOptionDataEntry {
         &self.0
     }
 }
+
 #[cfg(test)]
 mod tests {
     use crate::entities::FieldType;
     use crate::services::cell::CellDataOperation;
-    use crate::services::field::select_option::*;
+    
+    use crate::services::field::type_options::*;
     use crate::services::field::FieldBuilder;
-    use crate::services::field::{MultiSelectTypeOption, MultiSelectTypeOptionBuilder};
     use flowy_grid_data_model::revision::FieldRevision;
 
     #[test]
-    fn multi_select_test() {
+    fn single_select_test() {
         let google_option = SelectOption::new("Google");
         let facebook_option = SelectOption::new("Facebook");
         let twitter_option = SelectOption::new("Twitter");
-        let multi_select = MultiSelectTypeOptionBuilder::default()
+        let single_select = SingleSelectTypeOptionBuilder::default()
             .option(google_option.clone())
             .option(facebook_option.clone())
             .option(twitter_option);
 
-        let field_rev = FieldBuilder::new(multi_select)
+        let field_rev = FieldBuilder::new(single_select)
             .name("Platform")
             .visibility(true)
             .build();
 
-        let type_option = MultiSelectTypeOption::from(&field_rev);
+        let type_option = SingleSelectTypeOption::from(&field_rev);
 
-        let option_ids = vec![google_option.id.clone(), facebook_option.id.clone()].join(SELECTION_IDS_SEPARATOR);
+        let option_ids = vec![google_option.id.clone(), facebook_option.id].join(SELECTION_IDS_SEPARATOR);
         let data = SelectOptionCellChangeset::from_insert(&option_ids).to_str();
         let cell_data = type_option.apply_changeset(data.into(), None).unwrap();
-        assert_multi_select_options(
-            cell_data,
-            &type_option,
-            &field_rev,
-            vec![google_option.clone(), facebook_option],
-        );
+        assert_single_select_options(cell_data, &type_option, &field_rev, vec![google_option.clone()]);
 
         let data = SelectOptionCellChangeset::from_insert(&google_option.id).to_str();
         let cell_data = type_option.apply_changeset(data.into(), None).unwrap();
-        assert_multi_select_options(cell_data, &type_option, &field_rev, vec![google_option]);
+        assert_single_select_options(cell_data, &type_option, &field_rev, vec![google_option]);
 
         // Invalid option id
         let cell_data = type_option
             .apply_changeset(SelectOptionCellChangeset::from_insert("").to_str().into(), None)
             .unwrap();
-        assert_multi_select_options(cell_data, &type_option, &field_rev, vec![]);
+        assert_single_select_options(cell_data, &type_option, &field_rev, vec![]);
 
         // Invalid option id
         let cell_data = type_option
-            .apply_changeset(SelectOptionCellChangeset::from_insert("123,456").to_str().into(), None)
+            .apply_changeset(SelectOptionCellChangeset::from_insert("123").to_str().into(), None)
             .unwrap();
-        assert_multi_select_options(cell_data, &type_option, &field_rev, vec![]);
+
+        assert_single_select_options(cell_data, &type_option, &field_rev, vec![]);
 
         // Invalid changeset
         assert!(type_option.apply_changeset("123".to_owned().into(), None).is_err());
     }
 
-    fn assert_multi_select_options(
+    fn assert_single_select_options(
         cell_data: String,
-        type_option: &MultiSelectTypeOption,
+        type_option: &SingleSelectTypeOption,
         field_rev: &FieldRevision,
         expected: Vec<SelectOption>,
     ) {
