+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/text_type_option/mod.rs
@@ -1,0 +1,2 @@
+mod text_option;
+pub use text_option::*;
