+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/url_type_option/mod.rs
@@ -1,0 +1,6 @@
+mod tests;
+mod url_option;
+mod url_option_entities;
+
+pub use url_option::*;
+pub use url_option_entities::*;
