+++ frontend/rust-lib/flowy-grid/src/services/field/type_options/util/mod.rs
@@ -1,4 +1,3 @@
 mod cell_data_util;
 
-pub use crate::services::field::select_option::*;
 pub use cell_data_util::*;
