+++ frontend/rust-lib/flowy-grid/src/services/row/row_builder.rs
@@ -1,5 +1,5 @@
 use crate::services::cell::apply_cell_data_changeset;
-use crate::services::field::select_option::SelectOptionCellChangeset;
+use crate::services::field::SelectOptionCellChangeset;
 use flowy_error::{FlowyError, FlowyResult};
 use flowy_grid_data_model::revision::{gen_row_id, CellRevision, FieldRevision, RowRevision, DEFAULT_ROW_HEIGHT};
 use indexmap::IndexMap;
